+++ awc/src/error.rs
@@ -1,6 +1,7 @@
 //! Http client errors
 pub use actix_http::client::{ConnectError, InvalidUrl, SendRequestError};
 pub use actix_http::error::PayloadError;
+pub use actix_http::ws::HandshakeError as WsHandshakeError;
 pub use actix_http::ws::ProtocolError as WsProtocolError;
 
 use actix_http::http::{header::HeaderValue, Error as HttpError, StatusCode};