+++ awc/CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## [0.1.0-alpha.2] - 2019-04-xx
+## [0.1.0-alpha.2] - 2019-03-xx
 
 ### Added
 
@@ -11,6 +11,10 @@
 * Session wide headers.
 
 * Session wide basic and bearer auth.
+
+### Changed
+
+* Export `ws` sub-module with websockets related types
 
 
 ## [0.1.0-alpha.1] - 2019-03-28